+++ src/types.ts
@@ -1,21 +1,21 @@
 /**
  * Arguments passed to a mutation function
  * @template Source - The source type being transmuted from
- * @template Extra - Type of additional data for transmutation
+ * @template Context - Type of additional data for transmutation
  */
-export type TransmuterArgs<Source, Extra> = Extra extends undefined
+export type TransmuterArgs<Source, Context> = Context extends undefined
   ? { source: Source }
-  : { source: Source; extra: Extra }
+  : { source: Source; context: Context }
 
 /**
  * Function that performs a custom transmutation on a source object
  * @template Source - The source type being transmuted from
  * @template Target - The target type being transmuted to
  * @template TargetKey - The specific key of the target property being set
- * @template Extra - Type of additional data for transmutation
+ * @template Context - Type of additional data for transmutation
  */
-export type Transmuter<Source, Target, TargetKey extends keyof Target, Extra = undefined> =
-  (args: TransmuterArgs<Source, Extra>) => Target[TargetKey]
+export type Transmuter<Source, Target, TargetKey extends keyof Target, Context = undefined> =
+  (args: TransmuterArgs<Source, Context>) => Target[TargetKey]
 
 /**
  * Get keys of Source that have values assignable to Target[TargetKey]
@@ -28,11 +28,11 @@ type AssignableKeys<Source, Target, TargetKey extends keyof Target> = {
  * Defines how a property should be transmuted from source to target type
  * @template Source - The source type being transmuted from
  * @template Target - The target type being transmuted to
- * @template Extra - Type of additional data for transmutation
+ * @template Context - Type of additional data for transmutation
  */
-export type Schema<Source, Target, Extra = undefined> = {
+export type Schema<Source, Target, Context = undefined> = {
   [TargetKey in keyof Target]: {
     to: TargetKey
-    from: AssignableKeys<Source, Target, TargetKey> | Transmuter<Source, Target, TargetKey, Extra>
+    from: AssignableKeys<Source, Target, TargetKey> | Transmuter<Source, Target, TargetKey, Context>
   }
 }[keyof Target]